[C] Break up Editor component
