+++ client/src/global/components/form/ContentEditor/contexts/htmlBreadcrumbsContext.js
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+// This is here to prevent a circular dependency between Wrapper/Editor/Element
+export const HtmlBreadcrumbsContext = createContext([]);
