+++ client/src/global/components/form/ContentEditor/index.js
@@ -1,52 +1,15 @@
-import React, { createContext, useState } from "react";
-import PropTypes from "prop-types";
+import React from "react";
 import setter from "../setter";
-import FieldWrapper from "../FieldWrapper";
 import Loadable from "@docusaurus/react-loadable";
 
-export const HtmlBreadcrumbsContext = createContext([]);
-
 const Editor = Loadable({
-  loader: () => import("./components/Editor").then(editor => editor.default),
+  loader: () => import("./components/Wrapper").then(editor => editor.default),
   render(EditorComponent, props) {
     return <EditorComponent {...props} />;
   },
   loading: () => null
 });
 
-function ContentEditor({
-  set,
-  stylesheets,
-  initialHtmlValue,
-  initialSlateValue,
-  ...props
-}) {
-  const [selectedCrumb, setSelectedCrumb] = useState();
+export default setter(Editor);
 
-  return (
-    <FieldWrapper className="wide">
-      <HtmlBreadcrumbsContext.Provider
-        value={{ selectedCrumb, setSelectedCrumb }}
-      >
-        <Editor
-          set={set}
-          initialSlateValue={initialSlateValue}
-          initialHtmlValue={initialHtmlValue}
-          stylesheets={stylesheets}
-          {...props}
-        />
-      </HtmlBreadcrumbsContext.Provider>
-    </FieldWrapper>
-  );
-}
-
-export default setter(ContentEditor);
-
-ContentEditor.displayName = "Global.Form.ContentEditor";
-
-ContentEditor.propTypes = {
-  set: PropTypes.func,
-  stylesheets: PropTypes.arrayOf(PropTypes.object),
-  sectionId: PropTypes.string,
-  sectionBody: PropTypes.string
-};
+Editor.displayName = "Global.Form.ContentEditor.Loader";
